% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{add_tbl_prefix}
\alias{add_tbl_prefix}
\title{Add database storing schema to table names}
\usage{
add_tbl_prefix(tbl_names, id = "deidentified", status = "clean")
}
\arguments{
\item{tbl_names}{table name}

\item{id}{identified/hashed/deidentified}

\item{status}{raw/clean/generated}
}
\value{
a table in database with the apporpriate schemas
}
\description{
Add database storing schema to table names
}
