% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{all_equal_across_row}
\alias{all_equal_across_row}
\title{When the same data is stored in multiple tables/tibbles and you want to merge them together
use this function to check if COLUMN NAMES and COLUMN TYPES match}
\usage{
all_equal_across_row(df)
}
\arguments{
\item{df}{a data frame, can contain columns names across a list of tibbles or column types across a list of tibbles}
}
\value{
TRUE if column names or column types match, depends on what is being tested
}
\description{
When the same data is stored in multiple tables/tibbles and you want to merge them together
use this function to check if COLUMN NAMES and COLUMN TYPES match
}
