% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{clean_deidentified}
\alias{clean_deidentified}
\title{Function to create a clean table from raw table in database. It takes a database connection,
strings to find matching pattern `{id}${status}${tbl_name}`, range of years to filter applicantions by,
and using the first application for each applicant.}
\usage{
clean_deidentified(conn, tbl_name, min_year = 2013, max_year = 2017,
  id = "deidentified", status = "raw")
}
\arguments{
\item{conn}{database connection object}

\item{tbl_name}{table name, shold be deidentified}

\item{min_year}{minimum application year}

\item{max_year}{maximum application year}

\item{id}{table id from database schema}

\item{status}{table status from database schema}
}
\value{
a clean deidentified table for analysis
}
\description{
Function to create a clean table from raw table in database. It takes a database connection,
strings to find matching pattern `{id}${status}${tbl_name}`, range of years to filter applicantions by,
and using the first application for each applicant.
}
