% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{convert_or_retain}
\alias{convert_or_retain}
\title{convert_or_retain converts column variables to dates where possible
otherwise retain original column type}
\usage{
convert_or_retain(col)
}
\arguments{
\item{col}{column variable}
}
\value{
a date-time variable where possible, otherwise original column type
}
\description{
convert_or_retain converts column variables to dates where possible
otherwise retain original column type
}
