% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_histogram.R
\name{db_histogram}
\alias{db_histogram}
\title{db_histogram}
\usage{
db_histogram(data, bins = 30, output = "data", ...)
}
\arguments{
\item{data}{Data frame or Spark table}

\item{bins}{Number of bins for the Histogram. Defaults to 30.}

\item{output}{Sets the type of output, defaults to 'data'. Possible values: 'data', 'ggplot'}
}
\description{
db_histogram
}
\details{
Pushes the computation of the bins to the server and returns a data frame with the specified number of bins
}

