% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_raster.R
\name{db_raster}
\alias{db_raster}
\title{db_raster}
\usage{
db_raster(data, resolution = 300, output = "data", ...)
}
\arguments{
\item{data}{Data frame or Spark table}

\item{resolution}{Number of squares to return to plot as raster. Defaults to 300.}

\item{output}{Sets the type of output, defaults to 'data'. Possible values: 'data', 'ggplot'}
}
\description{
db_raster
}
\details{
Pushes the computation of the raster squares to the server and returns a data frame
}

