% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{deidentify_hashed}
\alias{deidentify_hashed}
\title{Function to build a de-identified table from a hashed table in database.
It takes a database connection and strings to find table matching pattern `{id}${status}${tbl_name}` and
characters list conatining names of fields to be dropped and returns a deidentified table}
\usage{
deidentify_hashed(conn, tbl_name, ..., id = "hashed", status = "raw")
}
\arguments{
\item{conn}{database connection}

\item{tbl_name}{table name in database}

\item{...}{variables names to be dropped}

\item{id}{table id from database schema}

\item{status}{table status from database schema}
}
\value{
a deidentified dataset that can be written to the table
}
\description{
Function to build a de-identified table from a hashed table in database.
It takes a database connection and strings to find table matching pattern `{id}${status}${tbl_name}` and
characters list conatining names of fields to be dropped and returns a deidentified table
}
