% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connections_functions.R
\name{disconnect_all}
\alias{disconnect_all}
\title{Severe all active MySQL connection}
\usage{
disconnect_all()
}
\value{
a logical list for each database connection that's severed
}
\description{
Severe all active MySQL connection
}
