% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{drop_cols_from_list}
\alias{drop_cols_from_list}
\title{Drop only columns listed as drop for each data table mentioned in yaml}
\usage{
drop_cols_from_list(df_list, cols_to_drop)
}
\arguments{
\item{df_list}{list of tibbles}

\item{cols_to_drop}{list of cols to drop for each dataframe}
}
\value{
a list of tibbles with formatted column names
}
\description{
Drop only columns listed as drop for each data table mentioned in yaml
}
