% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{drop_empty_cols}
\alias{drop_empty_cols}
\title{Drop columns in tibbles that have only NA values}
\usage{
drop_empty_cols(df)
}
\arguments{
\item{df}{a tibble to check of NA in columns}
}
\value{
a tibble with only non-empty columnn
}
\description{
Drop columns in tibbles that have only NA values
}
