% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{filter_first_app_only}
\alias{filter_first_app_only}
\title{Function to get the earliest application in the specified application year range}
\usage{
filter_first_app_only(df_list, min_year, max_year)
}
\arguments{
\item{df_list}{list of dataframes}

\item{min_year}{min application year}

\item{max_year}{max application year}
}
\value{
the earlier application for each application year
}
\description{
Function to get the earliest application in the specified application year range
}
