% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{find_tbls_with_col}
\alias{find_tbls_with_col}
\title{Find tibble with specified column name}
\usage{
find_tbls_with_col(tbl_list, col_name, conn)
}
\arguments{
\item{tbl_list}{a list of table names}

\item{col_name}{column name in tibble}

\item{conn}{database connection object}
}
\value{
return table that contain the specified table name
}
\description{
Find tibble with specified column name
}
