% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{fix_colnames}
\alias{fix_colnames}
\title{Fix all column names to standard format across tibbles}
\usage{
fix_colnames(df_list)
}
\arguments{
\item{df_list}{list of tibbles}
}
\value{
a list of tibles with formatted column names
}
\description{
Fix all column names to standard format across tibbles
}
