% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connections_functions.R
\name{get_mysql_conn}
\alias{get_mysql_conn}
\title{get_mysql_connection takes the input path to database credentials file stored in secure location
and returns database connection object to mysql database.}
\usage{
get_mysql_conn(path = "~", credentials_file = ".my.cnf", group = "rs-dbi")
}
\arguments{
\item{path}{path to credentials file}

\item{credentials_file}{the name of the credentials file}

\item{group}{client group - eud_db_owner, edu_db_read, edu_db_write}
}
\value{
MySQL connection object
}
\description{
get_mysql_connection takes the input path to database credentials file stored in secure location
and returns database connection object to mysql database.
}
