% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{get_tbl_names_by_stage}
\alias{get_tbl_names_by_stage}
\title{Get tibble name by stage and status}
\usage{
get_tbl_names_by_stage(conn, id, status)
}
\arguments{
\item{conn}{database connection object}

\item{id}{id from database naming schema}

\item{status}{status from database naming schema}
}
\value{
a list of tibble that matches id and staus
}
\description{
Get tibble name by stage and status
}
