% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{get_tbl_suffix}
\alias{get_tbl_suffix}
\title{Get the suffix of tibble names}
\usage{
get_tbl_suffix(tbl_names)
}
\arguments{
\item{tbl_names}{names of tiblles}
}
\value{
return suffix of tibble names
}
\description{
Get the suffix of tibble names
}
