% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{get_tbls_by_pattern}
\alias{get_tbls_by_pattern}
\title{Just like get tbl_by_pattern expect returns a named list of tibbles}
\usage{
get_tbls_by_pattern(conn, pattern, in_memory = TRUE)
}
\arguments{
\item{conn}{database connection object}

\item{pattern}{regular expression patterm for matching with table names}

\item{in_memory}{if true, put table into memory, if false, do lazy evaluation}
}
\value{
a list of named tibbles or tables
}
\description{
Just like get tbl_by_pattern expect returns a named list of tibbles
}
