% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{interpolate_and_execute}
\alias{interpolate_and_execute}
\title{Build and execute SQL query}
\usage{
interpolate_and_execute(query, conn, commit = FALSE, ...)
}
\arguments{
\item{query}{SQL query}

\item{conn}{database connection object}

\item{commit}{Boolean value, TRUE to explicitly write to database FALSE to not}

\item{...}{Additional paramaters, for passing in skeleton query for SQL query above}
}
\value{
Executed SQL query on the datbase, changes the databse in commit is TRUE
}
\description{
Build and execute SQL query
}
