% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{keep_cols_from_list}
\alias{keep_cols_from_list}
\title{Keep only columns listed as keep for each data table mentioned in yaml}
\usage{
keep_cols_from_list(df_list, cols_to_keep)
}
\arguments{
\item{df_list}{list of tibbles}

\item{cols_to_keep}{list of cols to drop for each dataframe}
}
\value{
a list of tibbles with formatted column names
}
\description{
Keep only columns listed as keep for each data table mentioned in yaml
}
