% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{put_tbl_to_memory}
\alias{put_tbl_to_memory}
\title{put_tbl_to_memory pulls down a table into memory from database and where applies,
covert dates to the appropriate date format}
\usage{
put_tbl_to_memory(conn, tbl_name)
}
\arguments{
\item{conn}{database connection object}

\item{tbl_name}{name of table in MySQL database}
}
\value{
a tibble
}
\description{
put_tbl_to_memory pulls down a table into memory from database and where applies,
covert dates to the appropriate date format
}
