% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{select_by_keyword}
\alias{select_by_keyword}
\title{Helper function to prase_yaml_cols, either drops for keeps columns based on input from a yaml file}
\usage{
select_by_keyword(df_list, cols_list)
}
\arguments{
\item{df_list}{a list of tibbles}

\item{cols_list}{a list of length one containing many column names}
}
\value{

}
\description{
Helper function to prase_yaml_cols, either drops for keeps columns based on input from a yaml file
}
