% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_session_info.R
\name{unload_workspace}
\alias{unload_workspace}
\title{Function to unload packages from workspace}
\usage{
unload_workspace()
}
\value{
clean workspace
}
\description{
Function to unload packages from workspace
}
