% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{write_to_database}
\alias{write_to_database}
\title{Function to write a named list of tibbles to database}
\usage{
write_to_database(df_list, conn)
}
\arguments{
\item{df_list}{a list of tibbles}

\item{conn}{database connection object}
}
\value{
TRUE if written to database, FALSE otherwise
}
\description{
Function to write a named list of tibbles to database
}
