% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{write_to_database_single}
\alias{write_to_database_single}
\title{Function to write a single tibble to database}
\usage{
write_to_database_single(df, conn, tbl_name)
}
\arguments{
\item{df}{tibble}

\item{conn}{database connection object}

\item{tbl_name}{name for table in database}
}
\value{
TRUE if written to database, FALSE otherwise
}
\description{
Function to write a single tibble to database
}
