% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{write_to_file}
\alias{write_to_file}
\title{Write data to local machine or share drive}
\usage{
write_to_file(df, path, filename)
}
\arguments{
\item{df}{tibble}

\item{path}{path to directory where file will be saved}

\item{filename}{filename}
}
\description{
Write data to local machine or share drive
}
