% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{write_to_separate_files}
\alias{write_to_separate_files}
\title{Write to file with unique names for files written}
\usage{
write_to_separate_files(df_nested, output_dir, filename_template)
}
\arguments{
\item{df_nested}{nested data frame with a list column called data}

\item{output_dir}{path to write files to}

\item{filename_template}{}
}
\description{
Write to file with unique names for files written
}
