% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_histogram.R
\name{db_histogram}
\alias{db_histogram}
\title{Computes the bins of the histogram in the server}
\usage{
db_histogram(data, bins = 30, output = "ggplot", ...)
}
\arguments{
\item{data}{Data frame or Spark table}

\item{bins}{Number of bins for the Histogram. Defaults to 30.}

\item{output}{Sets the type of output, defaults to 'data'. Possible values: 'data', 'ggplot'}
}
\description{
Computes the bins of the histogram in the server
}
\details{
The function will use the first column in the data set.  Using the dplyr::select statement
prior to calling db_raster will ease adding or removing a plotting step.  The 'data' value
for the 'db' argument will return a data frame instead of a plot.
}
\examples{
\dontrun{
# This will return a 'ggplot' of a 30 bin histogram of the 'distance' field
flights \%>\%
filter(arr_delay < 100) \%>\%
  select(distance) \%>\%
  db_histogram()
}


}

