% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{keep_cols_from_list}
\alias{keep_cols_from_list}
\title{Read in from a textfile with names for columns to keep in tibbles}
\usage{
keep_cols_from_list(df_list, col_keep_path)
}
\arguments{
\item{df_list}{list of tibbles}

\item{col_keep_path}{name of file with columns to keep}
}
\value{
a list of tibbles with formatted column names
}
\description{
Read in from a textfile with names for columns to keep in tibbles
}
