% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_functions.R
\name{parse_yaml_cols}
\alias{parse_yaml_cols}
\title{Parse a yaml dictionary with columns to keep or drop in tibbles}
\usage{
parse_yaml_cols(df_list, col_list_path)
}
\arguments{
\item{df_list}{list of tibbles}

\item{col_list_path}{path to file with columns to keep or drop}
}
\value{
a list of tibbles with the appropriate columns removed/kept
}
\description{
Parse a yaml dictionary with columns to keep or drop in tibbles
}
